+++ src/titiler/application/titiler/application/main.py
@@ -10,7 +10,12 @@ from titiler.application import __version__ as titiler_version
 from titiler.application.settings import ApiSettings
 from titiler.core.dependencies import DatasetPathParams
 from titiler.core.errors import DEFAULT_STATUS_CODES, add_exception_handlers
-from titiler.core.factory import MultiBaseTilerFactory, TilerFactory, TMSFactory
+from titiler.core.factory import (
+    AlgorithmFactory,
+    MultiBaseTilerFactory,
+    TilerFactory,
+    TMSFactory,
+)
 from titiler.core.middleware import (
     CacheControlMiddleware,
     LoggerMiddleware,
@@ -45,7 +50,16 @@ api_settings = ApiSettings()
 
 app = FastAPI(
     title=api_settings.name,
-    description="A lightweight Cloud Optimized GeoTIFF tile server",
+    description="""A modern dynamic tile server built on top of FastAPI and Rasterio/GDAL.
+
+---
+
+**Documentation**: <a href="https://developmentseed.org/titiler/" target="_blank">https://developmentseed.org/titiler/</a>
+
+**Source Code**: <a href="https://github.com/developmentseed/titiler" target="_blank">https://github.com/developmentseed/titiler</a>
+
+---
+    """,
     version=titiler_version,
     root_path=api_settings.root_path,
 )
@@ -114,11 +128,15 @@ if not api_settings.disable_mosaic:
 ###############################################################################
 # TileMatrixSets endpoints
 tms = TMSFactory()
-app.include_router(tms.router, tags=["TileMatrixSets"])
+app.include_router(tms.router, tags=["Tiling Schemes"])
+
+###############################################################################
+# Algorithms endpoints
+algorithms = AlgorithmFactory()
+app.include_router(algorithms.router, tags=["Algorithms"])
 
 add_exception_handlers(app, DEFAULT_STATUS_CODES)
 add_exception_handlers(app, MOSAIC_STATUS_CODES)
-
 
 # Set all CORS enabled origins
 if api_settings.cors_origins:
@@ -156,7 +174,13 @@ if api_settings.lower_case_query_parameters:
     app.add_middleware(LowerCaseQueryStringMiddleware)
 
 
-@app.get("/healthz", description="Health Check", tags=["Health Check"])
+@app.get(
+    "/healthz",
+    description="Health Check.",
+    summary="Health Check.",
+    operation_id="healthCheck",
+    tags=["Health Check"],
+)
 def ping():
     """Health check."""
     return {"ping": "pong!"}